python/requirements.txt
opentelemetry.sdk==1.23.0
opentelemetry.api==1.23.0
requests==2.31.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
